Added missing factor of 2 for Tensor Spectrum.
In ModeSolver.cpp the function ModeSolver::Find_PPS_Tensor returned with
a prefactor of 4 compared to ModeSolver::Find_PPS_Scalar. However this
factor of 4 accounts for the difference between scalar and tensor in the
Bunch Davies vacuum initial conditions. Hence we need a further factor
of 2 for the Tensor Spectrum to account for the two possible
polarization states.